requirements.txt
pygame
paho-mqtt
websockets>=16.0
